cache total count of library items in page info only
